Safari-611.1.10/LayoutTests/fast/canvas/pattern-too-large-to-create-expected.txt

CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 268435456).
PASS: Saw exception.
